Move getImportSpecifiers to modules folder
