+++ src/transforms/v2-to-v3/modules/index.ts
@@ -1,5 +1,6 @@
 export * from "./addV3ClientModules";
 export * from "./getImportEqualsDeclaration";
+export * from "./getImportSpecifiers";
 export * from "./getV2GlobalNameFromModule";
 export * from "./hasRequire";
 export * from "./removeV2ClientModule";
